requirements.txt
acronym_tools @ git+https://github.com/abhaybd/acronym.git
boto3~=1.36.17
fastapi~=0.115.7
h5py~=3.12.1
numpy~=1.26.4
open3d~=0.18.0
openai~=1.64.0
pillow~=10.2.0
pydantic~=2.10.5
pyrender~=0.1.45
requests~=2.0.0
scipy~=1.14.1
tqdm~=4.67.1
trimesh~=4.5.3
types-boto3~=1.36.17
types-boto3-s3~=1.36.15
uvicorn~=0.34.0
scene-synthesizer~=1.13.1
